+++ app/tabs/maps.tsx
@@ -445,12 +445,14 @@ async function getMapsTabContents(tab: TabManagerTab, signal: AbortSignal): Prom
                 ConfigConstants.views.Maps.mapsTabModeToSectionIDs[mode][sectionIndex]!;
             return await getMapsTabContentsRows({
                 tab,
-                // BUG: This crashes the tab if an entry with invalid data is present.
                 keys: await Promise.all(
                     targetKeys
                         .slice(start, end)
                         .map(
-                            async (key: KeyData): Promise<KeyData> => ({ ...key, data: (await NBT.parse((await tab.db!.get(key.rawKey))!)) as KeyData["data"] })
+                            async (key: KeyData): Promise<KeyData> => ({
+                                ...key,
+                                data: (await NBT.parse((await tab.db!.get(key.rawKey))!).catch((): null => null)) as KeyData["data"],
+                            })
                         )
                 ),
                 // eslint-disable-next-line @typescript-eslint/restrict-template-expressions -- There is only one section atm, if another section is ever added, remove this disable comment.