+++ Changelog.md
@@ -113,6 +113,7 @@
 -   Fixed a bug where the error dialogs for opening files with the app would display the path as `[object Object]` instead of the actual path.
 -   Fixed a bug where the contents of entries with a data type of `NBT` or `custom/JSONNBT` could not be searched with the `contents` query in the "View Files" tab.
 -   Fixed a bug where if a search in the "View Files" tab that did not include an `nbt` or `contents` query was performed prior to performing a search with an `nbt` or `contents` query, without switching off of the tab in between the two searches, the NBT data for the entries would not be loaded, resulting in the `nbt` or `contents` query not matching any NBT search results.
+-   Fixed a bug where the "Maps" tab would crash if an entry with invalid data was present while in Async Mode.
 
 ## Performance Improvements